Expose a base activity to extend from.
